Fix selector not working when passing falsy value
The primary use for selectors is using an array index as key. This means
that the value `0` is likely to be used quite often. However, due to the
use of a falsy check instead of an undefined check, passing 0 will not
cause an update when the key is unset. This is demonstrated by a test.
This commit replaces the falsy check with an undefined check.