Fix incorrect argument type of createEffect
When an initial value is passed to createComputed, createRenderEffect or
createEffect, and the return type of the passed function doesn't
include undefined, the argument will never be undefined either.
Additionally, when no initial value is given, there is no reason to
restrict the return type to a defined value. The code can handle
undefined values just fine.

Some examples that fail to typecheck, but work in practice have been
added as test cases.

This commit adds overloads to the mentioned functions to remove
undefined from the argument type when an initial value is given, and
allow undefined in the return type otherwise.