Fix types of functions which can return undefined
The setter returned by `createSignal()` accepts an optional argument,
but returns a non-optional type. This commit adds a test case where this
results in incorrectly typing an undefined value as defined.

Similarly, `createContext()` creates a context with a default value of
undefined. However, when used in `useContext()`, it will return a
non-undefined type. This is possible because `lookup()` returns `any`.
Typescript incorrectly assumed that `any` equals `T` and forgets that it
could also be undefined. This commit works around that by encoding the
possible undefined in the `T` stored in `Context<>`.