refactor(store): improve store types
This PR focuses more on the performance and inference issues the store types have.
It is not meant to be a complete typing overhaul.

Added `bigint` to `NotWrappable`.

`Store` was quite complicated and often confusing as it contained properties specific to the Store implementation.
However, most internal functions use `StoreNode`. As such `Store<T>` has been repurposed to simply be `DeepReadonly<T>`.

`DeepReadonly` has been simplified.

The Setter type is mostly unchanged externally, except:
- Only `number` is allowed to index Array types (before, for example Array methods and properties were allowed).
- The fallback overload is now much better and correctly constrains types but requires type annotations (they can't be inferred).
- All functions passed via the setter have parameters typed readonly as they should(?) be.

I tried to optimize the Setter as much as possible, and I think it's better at the same number of overloads but don't have any definitive proof of that.

`produce` is now able to infer types correctly from the `setStore` call.
It currently uses a `DeepMutable` type to circumvent `DeepReadonly` being part of the inferred input; the DeepReadonly part shouldn't be inferred.

fixes #732