include test files in tsconfig.json so that IDEs will type check using the same settings as with source files
Without this change, test files are not in strict mode, while Solid's
source is designed to work in strict mode, so intellisense and type
checking within IDEs will not show the full picture. For example, in
signals.type-tests.ts, without this change there will be errors on many
of the @ts-expect-error comments because non-strict mode won't trigger
the errors we've designed the types to produce in strict mode.