use tsconfig.build.json files during build
This is a trick so that the build does not include test files (otherwise
the output structure is modified and build fails), while allowing IDEs
to still properly type check test code using the same settings as source
code because IDEs will use tsconfig.json files by default.

Project references would be better, but that change is a *lot* more
involved. This works for now.