include test files in tsconfig.json so that IDEs will type check using the same settings as with source files (#784)
* include test files in tsconfig.json so that IDEs will type check using the same settings as with source files

Without this change, test files are not in strict mode, while Solid's
source is designed to work in strict mode, so intellisense and type
checking within IDEs will not show the full picture. For example, in
signals.type-tests.ts, without this change there will be errors on many
of the @ts-expect-error comments because non-strict mode won't trigger
the errors we've designed the types to produce in strict mode.

* use tsconfig.build.json files during build

This is a trick so that the build does not include test files (otherwise
the output structure is modified and build fails), while allowing IDEs
to still properly type check test code using the same settings as source
code because IDEs will use tsconfig.json files by default.

Project references would be better, but that change is a *lot* more
involved. This works for now.