refactor: remove DeepReadonly/Mutable from stores
The previous discussion on discord preferred more accurate types (readonly).
It may be time to revisit that decision.

Although typing stores as deeply readonly allows typescript to catch attempts at mutating stores, it has its own issues:
- Poor interoperability with existing code which may not be typed readonly despite being immutable
- Confusing errors for users with less typescript experience
- Classes are incorrectly made readonly since typescript can't differentiate classes and objects; currently this has an awkward workaround (#888)

The latter is an issue not initially considered. Additionally it seems that getters in object literals passed straight to createStore are unable to use `this`: https://playground.solidjs.com/?hash=-1721851158&version=1.4.1 (not sure if fixable -- not going to try unless this PR doesn't go through)

This PR removes all uses of `DeepReadonly` and `DeepMutable` and marks them deprecated.

It also:
- Types `unwrap` based on the item being unwrapped.
- Types `isWrappable` as a type guard.
based on #892.