Fix JSX runtime
This practically fixes the JSX development runtime, which passes several
other parameters, which currently results in Solid not rendering nodes.

It also fixes a case where a key parameter might be passed as `undefined`.
This does not happen normally when using a compiler that compiles JSX to
function calls, because Solid as far as I understand does not support keys.
However, as the key is defined in the API of `jsx`, I think it’s better to
cut it off.