Solid-Element: Add clarification on default prop values
It's been a while since I ran into this, but when I used solid element
in my project, I encountered the issue
where the components weren't reacting to
changes to props. It took quite a bit of head scratching
and delving into the code to see that only props
with default values are handled by the custom element.